docs(integration): address PR #393 review feedback
Move ommx_quantum_benchmarks_qaoa tutorial from usage/ to
integration/ and revise content per reviewer feedback. The main
substantive change is switching the QAOA optimizer to drive the
exact expectation value of the cost Hamiltonian via Aer's
EstimatorV2 instead of a shot estimate; this lets BFGS converge
without sampling noise (385 evals, expval 1.96, hit rate 4.6%).

Other changes:
- Rename title to use the dataset's package name ("OMMX Quantum
  Benchmarks") for clarity in both EN and JA.
- Add a one-line SCIP intro mentioning branch-and-bound up front,
  and normalize the JA wording to "分枝限定法".
- Drop the merit factor mention (introduced but never reused as a
  comparison metric).
- Replace "BaseDataset subclass" with "dataset class" /
  "データセットクラス".
- Remove the "Next steps:" section; the QAOA-math link already
  lives at the algorithm intro.
- JA polish: 体言止め headings switched to colons (目標, 最適値
  の到達, 集中度); "報告します" softened to "見てみます";
  "勝敗判定として読むものではありません" dropped.
- tests/docs/test_tutorials.py: drop the now-unneeded
  OPTIONAL_SKIP_MODULES entry, since integration/ is excluded
  from docs CI execution.
