+++ docs/en/integration/index.md
@@ -8,4 +8,5 @@ Notes on using Qamomile together with external libraries and quantum platforms.
 
 ## All articles
 
+- [Using OMMX Quantum Benchmarks (1): Implementing and Benchmarking Quantum Algorithms with Qamomile](ommx_quantum_benchmarks_qaoa) — Drive QAOA on a LABS instance loaded from the OMMX Quantum Benchmarks dataset and compare against SCIP
 - [qBraid Executor](qbraid_executor) — Run Qiskit circuits on qBraid-supported devices
