+++ docs/en/integration/ommx_quantum_benchmarks_qaoa.py
@@ -14,24 +14,23 @@
 
 # %% [markdown]
 # ---
-# tags: [usage, optimization, variational]
+# tags: [integration, optimization, variational]
 # ---
 #
-# # Using OMMX Benchmarks (1): Implementing and Benchmarking Quantum Algorithms with Qamomile
+# # Using OMMX Quantum Benchmarks (1): Implementing and Benchmarking Quantum Algorithms with Qamomile
 #
-# This tutorial shows how to drive a Qamomile-built quantum algorithm with
-# a problem taken straight from a public benchmark dataset, and how to
-# compare its solution quality against a classical solver in the same
-# end-to-end pipeline.
+# This tutorial shows how to run a Qamomile quantum algorithm on a problem
+# from a public benchmark dataset and compare its solution quality with a
+# classical solver in the same workflow.
 #
 # **Goal.** Build a QAOA solver in Qamomile, run it on a
 # **Low Autocorrelation Binary Sequences (LABS)** instance loaded from the
 # [OMMX Quantum Benchmarks](https://github.com/Jij-Inc/OmmxQuantumBenchmarks)
-# dataset, and benchmark the result against the classical SCIP solver
+# dataset, and compare the result with the classical SCIP solver
 # accessed through the [`ommx-pyscipopt-adapter`](https://github.com/Jij-Inc/ommx-pyscipopt-adapter).
 # Because both the QAOA path and the SCIP path consume the *same*
-# `ommx.v1.Instance`, the only thing that differs between them is the
-# algorithm itself — making this a clean apples-to-apples comparison.
+# `ommx.v1.Instance`, the main difference is the algorithm itself. This
+# makes the comparison direct.
 
 # %%
 # Install the additional packages used in this tutorial.
@@ -42,8 +41,8 @@
 #
 # **OMMX** ([Open Mathematical prograMming eXchange](https://jij-inc.github.io/ommx/en/introduction.html))
 # is a data format for exchanging mathematical optimization problems across
-# tools — its `ommx.v1.Instance` carries the objective, constraints,
-# decision-variable metadata, and (optionally) a reference solution.
+# tools. Its `ommx.v1.Instance` stores the objective, constraints,
+# decision-variable metadata, and optional reference solution.
 #
 # **OMMX Quantum Benchmarks** is a curated collection of optimization
 # benchmark instances distributed in this `ommx.v1.Instance` format. The
@@ -53,12 +52,12 @@
 # literature, including LABS, Market Split, Independent Set, and
 # Steiner-tree packing.
 #
-# Because each instance is just an `ommx.v1.Instance`, every Qamomile
-# entry point that already accepts `ommx.v1.Instance` — most importantly
-# `QAOAConverter` — can consume these benchmark problems with no extra
-# adapter code. The same `Instance` can also be fed to classical OMMX
-# adapters such as `ommx-pyscipopt-adapter`, which lets us reuse one
-# problem definition across quantum and classical workflows.
+# Because each benchmark instance is represented as an `ommx.v1.Instance`,
+# any Qamomile workflow that accepts `ommx.v1.Instance`, including
+# `QAOAConverter`, can consume these problems without extra adapter code.
+# The same `Instance` can also be passed to classical OMMX adapters such
+# as `ommx-pyscipopt-adapter`, so one problem definition can support both
+# quantum and classical workflows.
 
 # %% [markdown]
 # ## Problem: Low Autocorrelation Binary Sequences (LABS)
@@ -79,13 +78,12 @@
 # E(\boldsymbol{s}) = \sum_{k=1}^{n-1} c_k(\boldsymbol{s})^2,
 # $$
 #
-# which we want to *minimize* (equivalently: maximize the **merit factor**
-# $F = n^2 / (2 E)$). LABS is NP-hard and has long served as a stress test
-# for both classical and quantum heuristics.
+# which we want to *minimize*. LABS is NP-hard and has long served as a
+# stress test for both classical and quantum heuristics.
 #
 # ### Loading a LABS instance
 #
-# `Labs` exposes two models — `"integer"` (uses integer decision variables
+# `Labs` exposes two models: `"integer"` (uses integer decision variables
 # for $c_k$ plus the constraints that tie them to $\boldsymbol{s}$) and
 # `"quadratic_unconstrained"` (a QUBO reformulation that introduces
 # auxiliary binary variables $z_{i,k}$ encoding the products
@@ -106,8 +104,8 @@ print(f"First 5 instances: {dataset.available_instances['quadratic_unconstrained
 # uses $n + n(n-1) = 25$ binary variables (5 sequence bits
 # $x_i$ plus $n(n-1) = 20$ auxiliary $z_{i,k}$ bits). After
 # `Instance.to_qubo()` folds the penalty terms into the objective
-# and inactive variables are pruned, this lands at 15 qubits —
-# small enough to simulate locally, large enough that QAOA is
+# and inactive variables are pruned, this becomes a 15-qubit problem:
+# small enough to simulate locally, but large enough that QAOA is
 # non-trivial.
 
 # %%
@@ -121,15 +119,14 @@ print(f"Reference feasible: {reference_solution.feasible}")
 
 # %% [markdown]
 # The bundled reference solution gives the known optimum
-# $E^\star = 2$ for $n=5$ — equivalently, merit factor
-# $F^\star = 25 / (2 \cdot 2) = 6.25$. We will compare both
-# QAOA and SCIP against this value.
+# $E^\star = 2$ for $n=5$. We will compare both QAOA and SCIP
+# against this value.
 
 # %% [markdown]
 # ## Algorithm: QAOA
 #
-# Rather than reach for the high-level `QAOAConverter`, we build the
-# QAOA pipeline from scratch with `@qkernel`, following the recipe in
+# Rather than use the high-level `QAOAConverter`, we build the
+# QAOA workflow from scratch with `@qkernel`, following the recipe in
 # [QAOA for MaxCut: Building the Circuit from Scratch](../algorithm/qaoa_maxcut).
 # Refer to that tutorial for the gate-by-gate derivation; here we
 # focus on the implementation.
@@ -137,11 +134,11 @@ print(f"Reference feasible: {reference_solution.feasible}")
 # %% [markdown]
 # ### Spin model from the OMMX instance
 #
-# `Instance.to_qubo()` folds the (penalty-form) QUBO out of the
-# `ommx.v1.Instance`. We then wrap it in a `BinaryModel` and switch to
-# the spin (-1/+1) domain, which is what the QAOA cost layer expects.
-# We also normalize the coefficients so the cost-landscape scale stays
-# comparable across runs.
+# `Instance.to_qubo()` converts the penalty-form `ommx.v1.Instance` into
+# a QUBO. We then wrap it in a `BinaryModel` and switch to the spin
+# (-1/+1) domain, which is what the QAOA cost layer expects. We also
+# normalize the coefficients so the energy scale stays comparable across
+# runs.
 
 # %%
 import ommx.v1
@@ -161,10 +158,36 @@ spin_model = (
 print(f"QAOA qubits: {spin_model.num_bits}")
 
 # %% [markdown]
+# ### Cost Hamiltonian
+#
+# To drive the optimizer with the exact expectation value rather than a
+# shot estimate, we build the Ising cost Hamiltonian directly from the
+# spin-model coefficients: $Z_i$ terms for the linear part and
+# $Z_i Z_j$ terms for the quadratic part.
+
+# %%
+import qamomile.observable as qm_o
+
+cost_hamiltonian = qm_o.Hamiltonian()
+for i, hi in spin_model.linear.items():
+    if abs(hi) > 1e-12:
+        cost_hamiltonian.add_term((qm_o.PauliOperator(qm_o.Pauli.Z, i),), hi)
+for (i, j), Jij in spin_model.quad.items():
+    if abs(Jij) > 1e-12:
+        cost_hamiltonian.add_term(
+            (qm_o.PauliOperator(qm_o.Pauli.Z, i), qm_o.PauliOperator(qm_o.Pauli.Z, j)),
+            Jij,
+        )
+cost_hamiltonian.constant = spin_model.constant
+
+# %% [markdown]
 # ### QAOA qkernels
 #
-# Three small qkernels — uniform superposition, cost layer, mixer layer —
-# composed into the full ansatz.
+# The ansatz uses three small qkernels: a uniform superposition, a cost
+# layer, and a mixer layer. We compose them into a state-preparation
+# qkernel `qaoa_state`, then wrap it twice: once with `qmc.expval` to
+# get an expectation-value qkernel for the optimizer, and once with
+# `qmc.measure` to get a sampling qkernel for the final shot histogram.
 
 # %%
 import qamomile.circuit as qmc
@@ -204,7 +227,37 @@ def mixer_layer(
 
 
 @qmc.qkernel
-def qaoa_ansatz(
+def qaoa_state(
+    p: qmc.UInt,
+    quad: qmc.Dict[qmc.Tuple[qmc.UInt, qmc.UInt], qmc.Float],
+    linear: qmc.Dict[qmc.UInt, qmc.Float],
+    n: qmc.UInt,
+    gammas: qmc.Vector[qmc.Float],
+    betas: qmc.Vector[qmc.Float],
+) -> qmc.Vector[qmc.Qubit]:
+    q = superposition(n)
+    for layer in qmc.range(p):
+        q = cost_layer(quad, linear, q, gammas[layer])
+        q = mixer_layer(q, betas[layer])
+    return q
+
+
+@qmc.qkernel
+def qaoa_expval(
+    p: qmc.UInt,
+    quad: qmc.Dict[qmc.Tuple[qmc.UInt, qmc.UInt], qmc.Float],
+    linear: qmc.Dict[qmc.UInt, qmc.Float],
+    n: qmc.UInt,
+    gammas: qmc.Vector[qmc.Float],
+    betas: qmc.Vector[qmc.Float],
+    H: qmc.Observable,
+) -> qmc.Float:
+    q = qaoa_state(p, quad, linear, n, gammas, betas)
+    return qmc.expval(q, H)
+
+
+@qmc.qkernel
+def qaoa_sampling(
     p: qmc.UInt,
     quad: qmc.Dict[qmc.Tuple[qmc.UInt, qmc.UInt], qmc.Float],
     linear: qmc.Dict[qmc.UInt, qmc.Float],
@@ -212,19 +265,20 @@ def qaoa_ansatz(
     gammas: qmc.Vector[qmc.Float],
     betas: qmc.Vector[qmc.Float],
 ) -> qmc.Vector[qmc.Bit]:
-    q = superposition(n)
-    for layer in qmc.range(p):
-        q = cost_layer(quad, linear, q, gammas[layer])
-        q = mixer_layer(q, betas[layer])
+    q = qaoa_state(p, quad, linear, n, gammas, betas)
     return qmc.measure(q)
 
 
 # %% [markdown]
 # ### Transpile and optimize
 #
-# Transpile with $p = 3$ layers, then run COBYLA on the mean spin-model
-# energy of the sampled bitstrings. We seed both the AerSimulator and
-# NumPy so the trajectory is reproducible.
+# Transpile both kernels with $p = 3$ layers. The expectation-value
+# executable drives the optimizer; the sampling executable is used
+# later for the final shot histogram. We feed the optimizer the
+# *exact* expectation value of the cost Hamiltonian via Aer's
+# `EstimatorV2` primitive, which keeps the BFGS finite-difference
+# gradient free of sampling noise. We seed NumPy so the parameter
+# trajectory is reproducible.
 
 # %%
 import os
@@ -232,14 +286,27 @@ import time
 
 import numpy as np
 from qiskit_aer import AerSimulator
+from qiskit_aer.primitives import EstimatorV2
 from scipy.optimize import minimize
 
 from qamomile.qiskit import QiskitTranspiler
+from qamomile.qiskit.transpiler import QiskitExecutor
 
 p = 3
 transpiler = QiskitTranspiler()
-executable = transpiler.transpile(
-    qaoa_ansatz,
+expval_executable = transpiler.transpile(
+    qaoa_expval,
+    bindings={
+        "p": p,
+        "quad": spin_model.quad,
+        "linear": spin_model.linear,
+        "n": spin_model.num_bits,
+        "H": cost_hamiltonian,
+    },
+    parameters=["gammas", "betas"],
+)
+sampling_executable = transpiler.transpile(
+    qaoa_sampling,
     bindings={
         "p": p,
         "quad": spin_model.quad,
@@ -250,13 +317,13 @@ executable = transpiler.transpile(
 )
 
 SEED = 42
-executor = transpiler.executor(
-    backend=AerSimulator(seed_simulator=SEED, max_parallel_threads=1)
+executor = QiskitExecutor(
+    backend=AerSimulator(seed_simulator=SEED, max_parallel_threads=1),
+    estimator=EstimatorV2(),
 )
 
 docs_test_mode = os.environ.get("QAMOMILE_DOCS_TEST") == "1"
-sample_shots = 256 if docs_test_mode else 1024
-maxiter = 20 if docs_test_mode else 300
+maxiter = 5 if docs_test_mode else 50
 
 rng = np.random.default_rng(SEED)
 initial_params = rng.uniform(0, np.pi, 2 * p)
@@ -265,17 +332,14 @@ cost_history: list[float] = []
 
 
 def cost_fn(params: np.ndarray) -> float:
-    """Sample the QAOA circuit at `params` and return the mean spin-model energy."""
+    """Return the exact expectation value of the cost Hamiltonian at `params`."""
     gammas = list(params[:p])
     betas = list(params[p:])
-    job = executable.sample(
+    job = expval_executable.run(
         executor,
-        shots=sample_shots,
         bindings={"gammas": gammas, "betas": betas},
     )
-    result = job.result()
-    decoded = spin_model.decode_from_sampleresult(result)
-    energy = decoded.energy_mean()
+    energy = job.result()
     cost_history.append(energy)
     return energy
 
@@ -284,14 +348,14 @@ t0 = time.perf_counter()
 res = minimize(
     cost_fn,
     initial_params,
-    method="COBYLA",
+    method="BFGS",
     options={"maxiter": maxiter},
 )
 qaoa_optimize_time = time.perf_counter() - t0
 
-print(f"Optimized mean energy (normalized): {res.fun:.4f}")
-print(f"Function evaluations:               {res.nfev}")
-print(f"Wall time:                          {qaoa_optimize_time:.2f} s")
+print(f"Optimized expectation value (normalized): {res.fun:.4f}")
+print(f"Function evaluations:                     {res.nfev}")
+print(f"Wall time:                                {qaoa_optimize_time:.2f} s")
 
 # %%
 import matplotlib.pyplot as plt
@@ -299,7 +363,7 @@ import matplotlib.pyplot as plt
 plt.figure(figsize=(8, 4))
 plt.plot(cost_history, color="#2696EB")
 plt.xlabel("Iteration")
-plt.ylabel("Mean energy (normalized)")
+plt.ylabel("Expectation value (normalized)")
 plt.title("QAOA Optimization Progress (LABS, n=5)")
 plt.show()
 
@@ -307,14 +371,14 @@ plt.show()
 # ### Final sampling
 #
 # We sample once more with the optimized parameters and a larger shot
-# count, then decode against the original `ommx.v1.Instance` so the
+# count. Then we decode against the original `ommx.v1.Instance`, so the
 # returned `ommx.v1.SampleSet` reports the original QUBO objective
-# directly. For this QUBO formulation, samples whose auxiliary $z$
-# variables correctly encode the products $x_i x_{i+k+1}$ incur zero
-# penalty and the objective equals the true LABS energy
-# $E(\boldsymbol{s}) = \sum_k c_k^2$; samples that violate the
+# directly. In this QUBO formulation, samples whose auxiliary $z$
+# variables correctly encode the products $x_i x_{i+k+1}$ incur no
+# penalty, and the objective equals the LABS energy
+# $E(\boldsymbol{s}) = \sum_k c_k^2$. Samples that violate the
 # implicit $z = x_i x_{i+k+1}$ relation pay an additive penalty
-# proportional to the placeholder $P$ baked into the instance.
+# proportional to the placeholder $P$ built into the instance.
 
 # %%
 def evaluate_with_ommx(
@@ -342,7 +406,7 @@ gammas_opt = list(res.x[:p])
 betas_opt = list(res.x[p:])
 final_shots = 256 if docs_test_mode else 4096
 
-final_result = executable.sample(
+final_result = sampling_executable.sample(
     executor,
     shots=final_shots,
     bindings={"gammas": gammas_opt, "betas": betas_opt},
@@ -364,10 +428,10 @@ print(f"Reference E*:         {ref_E}")
 # QAOA returns a distribution over bitstrings, not a single answer. The
 # histogram below shows the QUBO objective of every shot at the
 # optimized parameters. The red dashed line marks the reference
-# optimum $E^\star$. Samples that fall on (or just to the right of) that
-# line are the ones whose $x$'s actually minimize the LABS sum and
-# whose $z$'s correctly encode the products; samples far to the right
-# are paying penalty for inconsistent $z$'s.
+# optimum $E^\star$. Samples on, or just to the right of, that line have
+# $x$ values that minimize the LABS sum and $z$ values that correctly
+# encode the products. Samples far to the right pay a penalty for
+# inconsistent $z$ values.
 
 # %%
 objectives = qaoa_summary["objective"].to_numpy()
@@ -389,11 +453,13 @@ plt.show()
 # %% [markdown]
 # ## Classical baseline: SCIP via the OMMX adapter
 #
-# The same `ommx.v1.Instance` is consumed by
+# **SCIP** is a branch-and-bound-based MILP/QUBO solver that can certify
+# an optimum when the solve completes. The same `ommx.v1.Instance`
+# is consumed by
 # `ommx_pyscipopt_adapter.OMMXPySCIPOptAdapter.solve`, which hands the
-# problem to the SCIP MILP/QUBO solver via PySCIPOpt and returns an
-# `ommx.v1.Solution` evaluated against the *original* instance — so its
-# `.objective` is directly comparable to QAOA's.
+# problem to SCIP via PySCIPOpt and returns an `ommx.v1.Solution`
+# evaluated against the *original* instance. Its `.objective` is
+# therefore directly comparable to QAOA's.
 
 # %%
 import ommx_pyscipopt_adapter
@@ -411,7 +477,7 @@ print(f"Wall time:    {scip_solve_time:.3f} s")
 # ## Results comparison
 #
 # SCIP returns a single optimum deterministically, while QAOA returns a
-# *distribution* over bitstrings — so we report QAOA's **best shot**
+# *distribution* over bitstrings. We report QAOA's **best shot**
 # (the lowest-objective bitstring seen across all samples) and its
 # **hit rate** on the reference optimum (the fraction of shots that
 # achieved $E^\star$).
@@ -435,18 +501,17 @@ print(f"QAOA hit rate on E* = {ref_E}: {hit_rate:.1%}  ({final_shots} shots)")
 #    optimum $E^\star = 2$, so QAOA is *capable* of finding the optimal
 #    sequence at $n = 5$ with only $p = 3$ layers.
 # 2. **Concentration.** QAOA's value lies in concentrating sampling
-#    probability on low-energy bitstrings — the hit rate above (and
-#    the left tail of the histogram) is the quantitative version of
+#    probability on low-energy bitstrings. The hit rate above, together
+#    with the left tail of the histogram, is the quantitative version of
 #    that statement.
 #
-# The wall-time column should be read for the *shape* of the comparison,
-# not as a head-to-head verdict: SCIP runs natively against the QUBO on
-# the CPU, while the QAOA timing covers the full classical-quantum
+# The wall-time column shows the *shape* of the comparison, not a
+# head-to-head verdict. SCIP runs directly against the QUBO on the CPU,
+# while the QAOA timing covers the full classical-quantum
 # optimization loop on a state-vector simulator. As $n$ grows, the
-# *qualitative* trade-off — exact-but-exponential branch-and-bound vs.
-# heuristic-but-polynomial-depth circuits — is what makes benchmark
-# datasets like OMMX Quantum Benchmarks useful for evaluating both
-# sides.
+# *qualitative* trade-off between exact branch-and-bound and
+# polynomial-depth heuristic circuits is what makes benchmark datasets
+# like OMMX Quantum Benchmarks useful for evaluating both sides.
 
 # %% [markdown]
 # ## Summary
@@ -457,7 +522,7 @@ print(f"QAOA hit rate on E* = {ref_E}: {hit_rate:.1%}  ({final_shots} shots)")
 #    dataset as an `ommx.v1.Instance`.
 # 2. Extracted the QUBO with `Instance.to_qubo()`, wrapped it in a
 #    `BinaryModel`, switched to the spin domain, and ran a hand-written
-#    QAOA ansatz (`@qkernel`) against it through
+#    QAOA ansatz (using `@qkernel`) against it through
 #    `QiskitTranspiler` + `AerSimulator`.
 # 3. Compared the QAOA output (best shot, hit rate, sampling
 #    distribution) against SCIP via
@@ -466,18 +531,9 @@ print(f"QAOA hit rate on E* = {ref_E}: {hit_rate:.1%}  ({final_shots} shots)")
 #
 # The pattern generalizes: any other QOBLIB dataset
 # (`Marketsplit`, `IndependentSet`, `Network`, …) plugs into the same
-# pipeline — load with the corresponding `BaseDataset` subclass,
-# extract the QUBO via `Instance.to_qubo()`, and reuse the same
-# `BinaryModel` + QAOA ansatz + transpile loop. Larger instances will
+# pipeline: load with the corresponding dataset class, extract the
+# QUBO via `Instance.to_qubo()`, and reuse the same `BinaryModel` +
+# QAOA ansatz + transpile loop. Larger instances will
 # eventually outgrow local simulators, at which point the same
 # `executable` can be re-targeted to other Qamomile backends
 # (`QuriPartsTranspiler`, `CudaqTranspiler`, …) or real hardware.
-#
-# **Next steps:**
-#
-# - For the QAOA mathematics itself, see
-#   [QAOA for MaxCut](../algorithm/qaoa_maxcut) and
-#   [QAOA for Graph Partitioning](../algorithm/qaoa_graph_partition).
-# - To swap in a different benchmark family, inspect
-#   `Labs().available_instances`, `IndependentSet().available_instances`,
-#   etc., from `ommx_quantum_benchmarks.qoblib`.
