+++ docs/en/usage/index.md
@@ -9,4 +9,3 @@ How-to guides for using individual Qamomile modules.
 ## All articles
 
 - [How to use BinaryModel](binary_model) — Build unconstrained binary/spin models from BinaryExpr, QUBO/HUBO/Ising, or OMMX
-- [Using OMMX Benchmarks (1): Implementing and Benchmarking Quantum Algorithms with Qamomile](ommx_quantum_benchmarks_qaoa) — Drive QAOA on a LABS instance loaded from the OMMX Quantum Benchmarks dataset and compare against SCIP
