+++ docs/ja/usage/index.md
@@ -9,4 +9,3 @@ Qamomileの個別モジュールを使うためのHow-toガイドです。
 ## すべての記事
 
 - [`BinaryModel`の使い方](binary_model) — 制約なしのバイナリ/スピン変数モデルをQUBO/HUBO/IsingあるいはOMMXから構築する
-- [OMMXベンチマークの活用 (1): Qamomileによる量子アルゴリズムの実装とベンチマーク](ommx_quantum_benchmarks_qaoa) — OMMX Quantum Benchmarksから取得したLABSインスタンス上でQAOAを動かし、SCIPと比較する
